backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
prophet==1.1.5
holidays==0.37
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.17.0
pycountry==24.6.1